modularized the IO. Finished the xml part. Working on ENSDF
